+++ packages/harmony/src/components/button/FilterButton/FilterButton.tsx
@@ -1,45 +1,39 @@
-import { forwardRef, RefObject, useRef, useState, useCallback } from 'react'
+import { forwardRef, useRef, useState, useCallback } from 'react'
 
 import { CSSObject, useTheme } from '@emotion/react'
 
 import { BaseButton } from 'components/button/BaseButton/BaseButton'
-import { TextInput, TextInputSize } from 'components/input'
-import { Flex, Paper } from 'components/layout'
-import { Popup } from 'components/popup'
 import { useControlled } from 'hooks/useControlled'
-import { IconCaretDown, IconCloseAlt, IconSearch } from 'icons'
+import { IconCaretDown, IconCloseAlt } from 'icons'
 
-import { FilterButtonOption, FilterButtonProps } from './types'
+import { FilterButtonProps } from './types'
 
 export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
   function FilterButton(props, ref) {
     const {
-      selection: selectionProp,
-      label,
-      options,
-      onSelect,
+      value: valueProp,
+      children,
+      label: labelProp,
+      onChange,
       disabled,
-      showFilterInput,
-      filterInputPlaceholder = 'Search',
       variant = 'fillContainer',
       size = 'default',
-      iconRight = IconCaretDown,
-      popupAnchorOrigin,
-      popupMaxHeight,
-      popupTransformOrigin,
-      popupPortalLocation,
-      popupZIndex
+      iconRight = IconCaretDown
     } = props
     const { color, cornerRadius, spacing, typography } = useTheme()
-    const [selection, setSelection] = useControlled({
-      controlledProp: selectionProp,
+    const [value, setValue] = useControlled({
+      controlledProp: valueProp,
       defaultValue: null,
-      stateName: 'selection',
+      stateName: 'value',
       componentName: 'FilterButton'
     })
-    const [filterInputValue, setFilterInputValue] = useState('')
-    const selectedOption = options.find((option) => option.value === selection)
-    const selectedLabel = selectedOption?.label ?? selectedOption?.value
+
+    const [label, setLabel] = useControlled({
+      controlledProp: labelProp,
+      defaultValue: null,
+      stateName: 'label',
+      componentName: 'FilterButton'
+    })
 
     const [isOpen, setIsOpen] = useState(false)
 
@@ -76,7 +70,7 @@ export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
     }
 
     const activeStyle =
-      variant !== 'fillContainer' || selection === null
+      variant !== 'fillContainer' || value === null
         ? {
             border: `1px solid ${color.border.strong}`,
             background: color.background.surface2
@@ -89,7 +83,7 @@ export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
       border: `1px solid ${color.border.strong}`,
       borderRadius: cornerRadius.s,
       color:
-        variant === 'fillContainer' && selection !== null
+        variant === 'fillContainer' && value !== null
           ? color.static.white
           : color.text.default,
       gap: spacing.xs,
@@ -110,60 +104,31 @@ export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
 
       ...(size === 'small' ? smallStyles : defaultStyles),
       ...(isOpen ? activeStyle : {}),
-      ...(variant === 'fillContainer' && selection !== null
+      ...(variant === 'fillContainer' && value !== null
         ? fillContainerStyles
         : {})
     }
 
     const iconCss = size === 'small' ? smallIconStyles : defaultIconStyles
 
-    // Popup Styles
-    const optionCss: CSSObject = {
-      background: 'transparent',
-      border: 'none',
-      color: color.text.default,
-      fontWeight: typography.weight.medium,
-      gap: spacing.s,
-      paddingLeft: spacing.m,
-      paddingRight: spacing.m,
-      paddingTop: spacing.s,
-      paddingBottom: spacing.s,
-      width: '100%',
-      borderRadius: cornerRadius.s,
-      justifyContent: 'flex-start',
-
-      '&:hover': {
-        transform: 'none',
-        backgroundColor: color.secondary.s300,
-        color: color.special.white
-      },
-
-      '&:active': {
-        transform: 'none'
-      }
-    }
-    const optionIconCss: CSSObject = {
-      width: spacing.unit4,
-      height: spacing.unit4
-    }
-
     const handleButtonClick = useCallback(() => {
-      setFilterInputValue('')
-      if (variant === 'fillContainer' && selection !== null) {
-        setSelection(null)
+      if (variant === 'fillContainer' && value !== null) {
+        setValue(null)
+        setLabel(null)
         // @ts-ignore
-        onSelect?.(null)
+        onChange?.(null)
       } else {
         setIsOpen((isOpen: boolean) => !isOpen)
       }
-    }, [selection, variant, setIsOpen, setSelection, onSelect])
+    }, [value, variant, setIsOpen, setValue, setLabel, onChange])
 
-    const handleOptionSelect = useCallback(
-      (option: FilterButtonOption) => {
-        setSelection(option.value)
-        onSelect?.(option.value)
+    const handleChange = useCallback(
+      (value: string, label: string) => {
+        setValue(value)
+        setLabel(label)
+        onChange?.(value)
       },
-      [onSelect, setSelection]
+      [onChange, setValue, setLabel]
     )
 
     const anchorRef = useRef<HTMLButtonElement>(null)
@@ -177,7 +142,7 @@ export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
         }}
         onClick={handleButtonClick}
         iconRight={
-          variant === 'fillContainer' && selection !== null
+          variant === 'fillContainer' && value !== null
             ? IconCloseAlt
             : iconRight
         }
@@ -185,69 +150,13 @@ export const FilterButton = forwardRef<HTMLButtonElement, FilterButtonProps>(
         aria-haspopup='listbox'
         aria-expanded={isOpen}
       >
-        {selectedLabel ?? label}
-        <Popup
-          anchorRef={(ref as RefObject<HTMLElement>) || anchorRef}
-          isVisible={isOpen}
-          onClose={() => setIsOpen(false)}
-          anchorOrigin={popupAnchorOrigin}
-          transformOrigin={popupTransformOrigin}
-          portalLocation={popupPortalLocation}
-          zIndex={popupZIndex}
-        >
-          <Paper mt='s' border='strong' shadow='far'>
-            <Flex
-              p='s'
-              direction='column'
-              alignItems='flex-start'
-              role='listbox'
-              aria-label={selectedLabel ?? label ?? props['aria-label']}
-              aria-activedescendant={selectedLabel}
-              css={{ maxHeight: popupMaxHeight, overflowY: 'auto' }}
-            >
-              <Flex direction='column' w='100%' gap='s'>
-                {showFilterInput ? (
-                  <TextInput
-                    placeholder={filterInputPlaceholder}
-                    label={filterInputPlaceholder}
-                    size={TextInputSize.SMALL}
-                    startIcon={IconSearch}
-                    onClick={(e) => {
-                      e.stopPropagation()
-                    }}
-                    onChange={(e) => {
-                      setFilterInputValue(e.target.value)
-                    }}
-                  />
-                ) : null}
-                {options
-                  .filter(({ label }) => {
-                    return (
-                      !filterInputValue ||
-                      label
-                        ?.toLowerCase()
-                        .includes(filterInputValue.toLowerCase())
-                    )
-                  })
-                  .map((option) => (
-                    <BaseButton
-                      key={option.value}
-                      iconLeft={option.icon}
-                      styles={{
-                        button: optionCss,
-                        icon: optionIconCss
-                      }}
-                      onClick={() => handleOptionSelect(option)}
-                      aria-label={option.label ?? option.value}
-                      role='option'
-                    >
-                      {option.label ?? option.value}
-                    </BaseButton>
-                  ))}
-              </Flex>
-            </Flex>
-          </Paper>
-        </Popup>
+        {label}
+        {children({
+          isOpen,
+          setIsOpen,
+          handleChange,
+          anchorRef
+        })}
       </BaseButton>
     )
   }
