+++ packages/harmony/src/components/button/FilterButton/types.ts
@@ -1,8 +1,10 @@
+import { ReactNode } from 'react'
+
 import { IconComponent } from 'components/icon'
-import { Origin } from 'components/popup/types'
 
 export type FilterButtonSize = 'default' | 'small'
 
+// TODO: is replaceLabel still needed?
 export type FilterButtonVariant = 'fillContainer' | 'replaceLabel'
 
 export type FilterButtonOption = {
@@ -14,12 +16,15 @@ export type FilterButtonOption = {
   icon?: IconComponent
 }
 
+type ChildrenProps = {
+  isOpen: boolean
+  setIsOpen: (isOpen: boolean) => void
+  handleChange: (value: string, label: string) => void
+  anchorRef: React.RefObject<HTMLButtonElement>
+}
+
 export type FilterButtonProps = {
-  /**
-   * Selection options
-   * e.g. { label: 'Option A', icon: IconRadar }
-   */
-  options: FilterButtonOption[]
+  children: (props: ChildrenProps) => ReactNode
 
   /**
    * The text that appears on the button component.
@@ -34,9 +39,9 @@ export type FilterButtonProps = {
   'aria-label'?: string
 
   /**
-   * The selected value
+   * The value
    */
-  selection?: string | null
+  value?: string | null
 
   /**
    * The button size
@@ -61,49 +66,12 @@ export type FilterButtonProps = {
   iconRight?: IconComponent
 
   /**
-   * What to do when an option is selected
+   * What to do when the value is changed
    */
-  onSelect?: (label: string) => void
+  onChange?: (value: string) => void
 
   /**
    * Whether interaction is disabled
    */
   disabled?: boolean
-
-  /**
-   * Popup anchor origin
-   * @default { horizontal: 'center', vertical: 'bottom' }
-   */
-  popupAnchorOrigin?: Origin
-
-  /**
-   * Popup max height
-   */
-  popupMaxHeight?: number
-
-  /**
-   * Popup transform origin
-   * @default { horizontal: 'center', vertical: 'top' }
-   */
-  popupTransformOrigin?: Origin
-
-  /**
-   * Popup portal location passed to the inner popup
-   */
-  popupPortalLocation?: HTMLElement
-
-  /**
-   * zIndex applied to the inner Popup component
-   */
-  popupZIndex?: number
-
-  /**
-   * Show a text input to filter the options
-   */
-  showFilterInput?: boolean
-
-  /**
-   * Placeholder text for the filter input
-   */
-  filterInputPlaceholder?: string
 }
