+++ packages/harmony/src/components/button/index.ts
@@ -7,5 +7,7 @@ export { FollowButton } from './FollowButton/FollowButton'
 export * from './FollowButton/types'
 export { FilterButton } from './FilterButton/FilterButton'
 export * from './FilterButton/types'
+export { OptionsFilterButton } from './OptionsFilterButton/OptionsFilterButton'
+export * from './OptionsFilterButton/types'
 export { IconButton, IconButtonProps } from './IconButton/IconButton'
 export { UnstyledButton } from './UnstyledButton'
