+++ packages/web/src/components/payment-method/PaymentMethod.tsx
@@ -12,6 +12,7 @@ import {
   IconCreditCard,
   IconDonate,
   IconTransaction,
+  OptionsFilterButton,
   Radio,
   RadioGroup,
   Text
@@ -104,8 +105,8 @@ export const PaymentMethod = ({
               zIndex={zIndex.ADD_FUNDS_VENDOR_SELECTION_DRAWER}
             />
           ) : (
-            <FilterButton
-              onSelect={handleSelectVendor}
+            <OptionsFilterButton
+              onChange={handleSelectVendor}
               selection={selectedVendor?.toString()}
               variant='replaceLabel'
               options={vendorOptions}
