+++ packages/web/src/components/upload/TrackPreview.tsx
@@ -20,6 +20,7 @@ import {
   IconFileWAV as iconFileWav,
   IconFileWEBM,
   IconTrash,
+  OptionsFilterButton,
   Text
 } from '@audius/harmony'
 import cn from 'classnames'
@@ -150,7 +151,7 @@ export const TrackPreview = (props: TrackPreviewProps) => {
       <Flex alignItems='center'>
         {isStem && onEditStemCategory ? (
           <Box mr='xl'>
-            <FilterButton
+            <OptionsFilterButton
               label={messages.selectType}
               options={stemCategories}
               popupAnchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
@@ -158,7 +159,7 @@ export const TrackPreview = (props: TrackPreviewProps) => {
                 vertical: 'top',
                 horizontal: 'right'
               }}
-              onSelect={(label) => onEditStemCategory(label as StemCategory)}
+              onChange={(label) => onEditStemCategory(label as StemCategory)}
               selection={stemCategory?.toString() ?? null}
               popupZIndex={zIndex.STEMS_AND_DOWNLOADS_FILTER_BUTTON_POPUP}
               disabled={!allowCategorySwitch}
