+++ packages/web/src/pages/dashboard-page/components/ArtistContentSection.tsx
@@ -8,7 +8,8 @@ import {
   Flex,
   FilterButton,
   TextInput,
-  TextInputSize
+  TextInputSize,
+  OptionsFilterButton
 } from '@audius/harmony'
 
 import { ArtistDashboardAlbumsTab } from './ArtistDashboardAlbumsTab'
@@ -109,9 +110,9 @@ export const ArtistContentSection = () => {
               onClick={() => onClickPill(Pills.ALBUMS)}
             />
           </Flex>
-          <FilterButton
-            onSelect={handleSelectFilter}
-            selection={isTracks ? selectedTrackFilter : selectedAlbumFilter}
+          <OptionsFilterButton
+            onChange={handleSelectFilter}
+            value={isTracks ? selectedTrackFilter : selectedAlbumFilter}
             label={messages.allReleases}
             options={filterButtonOptions}
             popupAnchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
